Revert "Default XMP_TYPE_LANG_ALT to "x-default" according to spec."
This reverts commit a4ccfcee72e13228ced79185d1116c846df63892, because
the encoder needs to generate XMP data with attributes and their
language code if they're set in the model. This commit was hiding the
problem, that the xmp_model_entry widgets changed the raw value in the
tree model and set a wrong type of value therefore overwriting important
data.